refactor: create new package fs in dynamic-node-id
