+++ zeebe/dynamic-node-id-provider/src/main/java/io/camunda/zeebe/dynamic/nodeid/fs/DirectoryInitializationInfo.java
@@ -5,7 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.dynamic.nodeid;
+package io.camunda.zeebe.dynamic.nodeid.fs;
+
+import io.camunda.zeebe.dynamic.nodeid.Version;
 
 /**
  * @param initializedAt epoch in milliseconds when the directory was initialized
