+++ zeebe/dynamic-node-id-provider/src/main/java/io/camunda/zeebe/dynamic/nodeid/fs/NodeIdBasedDataDirectoryProvider.java
@@ -5,8 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.dynamic.nodeid;
+package io.camunda.zeebe.dynamic.nodeid.fs;
 
+import io.camunda.zeebe.dynamic.nodeid.NodeInstance;
 import java.nio.file.Path;
 import java.util.concurrent.CompletableFuture;
