+++ zeebe/dynamic-node-id-provider/src/test/java/io/camunda/zeebe/dynamic/nodeid/fs/PreConfiguredDataDirectoryProviderTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.dynamic.nodeid;
+package io.camunda.zeebe.dynamic.nodeid.fs;
 
 import static org.assertj.core.api.Assertions.assertThat;
