+++ zeebe/dynamic-node-id-provider/src/test/java/io/camunda/zeebe/dynamic/nodeid/fs/VersionedNodeIdBasedDataDirectoryProviderTest.java
@@ -5,12 +5,14 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.dynamic.nodeid;
+package io.camunda.zeebe.dynamic.nodeid.fs;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.zeebe.dynamic.nodeid.NodeInstance;
+import io.camunda.zeebe.dynamic.nodeid.Version;
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
